ISSUE-10 Add javadoc/scaladoc for public classes
